+++ test/test_qiskit.py
@@ -7,7 +7,7 @@ from c3.qiskit.c3_exceptions import C3QiskitError
 from qiskit.quantum_info import Statevector
 from qiskit import transpile
 from qiskit.providers import BackendV1 as Backend
-from qiskit import execute
+from qiskit import execute, QuantumCircuit
 import pytest
 
 
@@ -22,7 +22,9 @@ def test_backends():
 
 @pytest.mark.unit
 @pytest.mark.qiskit
-@pytest.mark.parametrize("backend", ["c3_qasm_perfect_simulator"])
+@pytest.mark.parametrize(
+    "backend", ["c3_qasm_perfect_simulator", "c3_qasm_physics_simulator"]
+)
 def test_get_backend(backend):
     """Test get_backend() which returns the backend with matching name
 
@@ -62,13 +64,13 @@ def test_transpile(get_test_circuit, backend):  # noqa
 @pytest.mark.qiskit
 @pytest.mark.slow
 @pytest.mark.parametrize("backend", ["c3_qasm_perfect_simulator"])
-def test_get_result(get_6_qubit_circuit, backend, get_result_qiskit):  # noqa
-    """Test the counts from running a 6 qubit Circuit
+def test_get_result(get_3_qubit_circuit, backend, get_result_qiskit):  # noqa
+    """Test the counts from running a 3 qubit Circuit
 
     Parameters
     ----------
-    get_6_qubit_circuit : callable
-        pytest fixture for a 6 qubit circuit
+    get_3_qubit_circuit : callable
+        pytest fixture for a 3 qubit circuit
     backend : str
         name of the backend which is to be tested
     simulation_type: str
@@ -80,7 +82,7 @@ def test_get_result(get_6_qubit_circuit, backend, get_result_qiskit):  # noqa
     received_backend = c3_qiskit.get_backend(backend)
     received_backend.set_device_config("test/quickstart.hjson")
     received_backend.disable_flip_labels()
-    qc = get_6_qubit_circuit
+    qc = get_3_qubit_circuit
     job_sim = execute(qc, received_backend, shots=1000)
     result_sim = job_sim.result()
 
@@ -119,5 +121,41 @@ def test_get_exception(get_bad_circuit, backend):  # noqa
     received_backend.set_device_config("test/quickstart.hjson")
     qc = get_bad_circuit
 
+    with pytest.raises(C3QiskitError):
+        execute(qc, received_backend, shots=1000)
+
+
+def test_qiskit_physics():
+    """API test for qiskit physics simulation"""
+    c3_qiskit = C3Provider()
+    physics_backend = c3_qiskit.get_backend("c3_qasm_physics_simulator")
+    physics_backend.set_device_config("test/qiskit.cfg")
+    qc = QuantumCircuit(3, 3)
+    qc.x(0)
+    qc.cx(0, 1)
+    job_sim = execute(qc, physics_backend)
+    print(job_sim.result().get_counts())
+
+
+@pytest.mark.parametrize(
+    "backend",
+    [
+        ("c3_qasm_perfect_simulator", "test/quickstart.hjson"),
+        ("c3_qasm_physics_simulator", "test/qiskit.cfg"),
+    ],
+)
+def test_too_many_qubits(backend):
+    """Check that error is raised when circuit has more qubits than device
+
+    Parameters
+    ----------
+    backend : tuple
+        name and device config of the backend to be tested
+    """
+    c3_qiskit = C3Provider()
+    received_backend = c3_qiskit.get_backend(backend[0])
+    received_backend.set_device_config(backend[1])
+    qc = QuantumCircuit(4, 4)
+    qc.x(1)
     with pytest.raises(C3QiskitError):
         execute(qc, received_backend, shots=1000)